renamed component which handels Keycloak UI aspects
