+++ src/app/core/user/user-account/user-account.component.html
@@ -42,9 +42,9 @@
             [disabled]="passwordChangeDisabled"
           ></app-password-form>
 
-          <app-password-button
+          <app-account-page
             [disabled]="passwordChangeDisabled"
-          ></app-password-button>
+          ></app-account-page>
         </div>
       </div>
     </ng-template>
